+++ emc-dao/src/main/java/com/huak/mdc/vo/MeterCollectA.java
@@ -33,7 +33,7 @@ public class MeterCollectA  implements Serializable{
 
         private String tag;
 
-        private Byte coef;
+        private Double coef;
 
         private String formula;
 
@@ -43,7 +43,7 @@ public class MeterCollectA  implements Serializable{
 
         private String comId;
 
-        public MeterCollectA(String id, String code, String name, String serialNo, String energyTypeId, Byte isreal, Byte istotal, String unitId,String unitname, Byte unitType, Byte isauto, String depict, String tag, Byte coef, String formula, Byte isprestore, Byte isdelete, String comId) {
+        public MeterCollectA(String id, String code, String name, String serialNo, String energyTypeId, Byte isreal, Byte istotal, String unitId,String unitname, Byte unitType, Byte isauto, String depict, String tag, Double coef, String formula, Byte isprestore, Byte isdelete, String comId) {
             this.id = id;
             this.code = code;
             this.name = name;
@@ -166,11 +166,11 @@ public class MeterCollectA  implements Serializable{
             this.tag = tag == null ? null : tag.trim();
         }
 
-        public Byte getCoef() {
+        public Double getCoef() {
             return coef;
         }
 
-        public void setCoef(Byte coef) {
+        public void setCoef(Double coef) {
             this.coef = coef;
         }
